+++ src/messenger/messenger.ts
@@ -291,7 +291,7 @@ export class Messenger {
     }
 
     /**
-     * By default, the {@link BroadcastChannel} is unreffed. Call this function to change that.
+     * By default, the {@link BroadcastChannel} is unreffed. Use this method to change that.
      * When `true`, [`ref()`](https://nodejs.org/api/worker_threads.html#broadcastchannelref) will be called.
      * When `false`, [`unref()`](https://nodejs.org/api/worker_threads.html#broadcastchannelunref) will be called.
      */