+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - More info about `metaData` to **Streaming data between threads** section of README.
 - `Messenger.closed` getter.
+- `Service.setRef()` method.
 
 ### Fixed